+++ packages/orq-rc/src/models/components/evaluatorresponseragas.ts
@@ -50,8 +50,8 @@ export const EvaluatorResponseRagas$inboundSchema: z.ZodType<
 > = z.object({
   _id: z.string(),
   description: z.string(),
-  created: z.string().default("2026-07-17T10:58:23.757Z"),
-  updated: z.string().default("2026-07-17T10:58:23.757Z"),
+  created: z.string().default("2026-07-17T12:17:54.934Z"),
+  updated: z.string().default("2026-07-17T12:17:54.934Z"),
   updated_by_id: z.nullable(z.string()).optional(),
   guardrail_config: z.any().optional(),
   type: z.literal("ragas"),