+++ packages/orq-rc/RELEASES.md
@@ -10909,3 +10909,13 @@ Based on:
 - [typescript v4.12.0-rc.37] packages/orq-rc
 ### Releases
 - [NPM v4.12.0-rc.37] https://www.npmjs.com/package/@orq-ai/node/v/4.12.0-rc.37 - packages/orq-rc
+
+## 2026-07-17 12:20:30
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.790.2 (2.918.3) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.12.0-rc.38] packages/orq-rc
+### Releases
+- [NPM v4.12.0-rc.38] https://www.npmjs.com/package/@orq-ai/node/v/4.12.0-rc.38 - packages/orq-rc